feat(obtener_pob_incidencia): created a function to obtain the incidence pop and warnings
A new function has been created to obtain the population that will be used in the incidence calculation. Depending on the value of the 'poblacion' parameter, it obtains either the at-risk population or the population projections from DANE. If there isn't any at-risk population available for the selected year, it obtains the population projections and displays different warnings to users based on the type of population obtained. It includes the corresponding documentation, examples, and validations.

Ref: #172 & #146